HIVE-22525 Refactor HiveOpConverter
